doc_classifier/pyproject.toml
[project]
name = "doc-classifier"
version = "0.1.0"
description = "Auto-detect document classification levels from metadata, labels, and watermarks."
readme = "TECHNICAL_GUIDE.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "python-docx",
    "openpyxl",
    "python-pptx",
    "pikepdf",
    "pdfplumber",
    "pdf2image",
    "pytesseract",
    "Pillow",
    "lxml"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["doc_classifier*"]
exclude = ["pipeline_output*", "test_documents*"]
